Improve the yuv420p conversion
It is better to take the average Cb and Cr values of the 4 pixels
instead of just using the first one.

This might be a bit slower, but I hope that it doesn't make a lot
of impact. If it does (especially on a Raspberry Pi) I might
revert this commit, because old method also looked fine.